Task 20.2 corrected
